+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+- Add method to reenqueue scheduler in squashed migrations [#226](https://github.com/hlascelles/que-scheduler/pull/226)
+
 ## 4.0.1 (2020-11-29)
 
 - Move setting schedule hash directly to config block [#225](https://github.com/hlascelles/que-scheduler/pull/225)